sync_qemu_models_i386: Do not overwrite existing models
We don't change definitions of CPU models which were already included in
a libvirt release to maintain migration compatibility. Thus the script
can just skip existing models and save us from having to drop the
changes it would do to them.

+++ src/cpu_map/sync_qemu_models_i386.py
@@ -535,6 +535,10 @@ def main():
 
     for model in models:
         name = os.path.join(args.outdir, f"x86_{model['name']}.xml")
+        if os.path.isfile(name):
+            # Ignore existing models as CPU models in libvirt should never
+            # change once released.
+            continue
         with open(name, "wt") as f:
             output_model(f, model)
